refactor: Debug class, info changes
